PFR-5: Improve map generation
The new method generates a set of paths, where each path is generated like this:
1. Pick a random starting point close to the surface
2. Pick a point that is reachable from the last point
3. If it isn't too close to any already existing points, add it to the level and
   use it as the new "last point"
4. Repeat 2 and 3 until a point that would be placed above the goal radius is
   generated. (The point that would be placed above the goal radius isn't
   included in the final path.)

To evaluate if a point is reachable, an approximation is used, with various
simplifications:
Gravity is assumed to be constant from the last point. In other words, whatever
    gravity is at the last point, is assumed to apply globally. This makes
    calculating time of flight and maximum height much easier.
Jump acceleration is assumed to only be applied for a single tick. In reality it
    can sometimes be applied for two ticks (but that's a bug).
Run acceleration is disregarded. So it's technically possible to reach further
    "sideways" than the maximum reachable value used, but it's really difficult
    to successfully land on a platform when jumping that far sideways.